Move AudioQueue to its own file
